Update
Flask cambiado a mysql. Aún pendiente de pruebas para correrlo en el front-end
