+++ orm/server_taught.py
@@ -83,37 +83,9 @@ with app.app_context():
     
 #Routes
 
-@app.route('/students', methods=['POST','GET'])
-def route_add_student():
-    if request.method == 'GET':
-        return get_students()
-    elif request.method == 'POST':
-        student = request.get_json()
-        return insert_student(student)
-
 #@app.route('/students/delete/<student_id>', methods=['GET', 'DELETE'])
 #def route_delete_student(student_id):
 #    return delete_student(student_id)
-
-@app.route('/students/get/<student_name>', methods=['GET'])
-def route_get_students_by_name(student_name):
-    res = db.session.query(Estudiante.id,Usuario.username,Usuario.email,Usuario.password).filter(Estudiante.usuario_id==Usuario.id).filter(Usuario.username == student_name).first()
-    students = {"estudiante_id":res[0],"usuario_username":res[1],"usuario_email":res[2],"usuario_password":res[3]}
-    return jsonify(students)
-
-@app.route('/teachers/get/<teacher_name>', methods=['GET'])
-def route_get_teachers_by_name(teacher_name):
-    res = db.session.query(Profesor.id,Usuario.username,Usuario.email,Usuario.password).filter(Profesor.usuario_id==Usuario.id).filter(Usuario.username == teacher_name).first()
-    teacher = {"profesor_id":res[0],"usuario_username":res[1],"usuario_email":res[2],"usuario_password":res[3]}
-    return jsonify(teacher)
-
-@app.route('/teachers', methods=['GET','POST'])
-def route_add_teacher():
-    if request.method == 'GET':
-        return get_teachers()
-    elif request.method == 'POST':
-        teacher = request.get_json()
-        return insert_teacher(teacher)
 
 @app.route('/taught', methods=['GET','POST'])
 def route_add_get_clase():
@@ -129,15 +101,6 @@ def route_add_get_clase():
     elif request.method == 'POST':
         clase = request.get_json()
         return insert_clase(clase)
-
-@app.route('/teachers/<teacher_course>', methods=['GET'])
-def route_get_teacher_by_course(teacher_course):
-    res = db.session.query(Profesor.id,Profesor.experiencia,Profesor.curso,Usuario.username,Usuario.email).filter(Profesor.usuario_id==Usuario.id).filter(Profesor.curso == teacher_course).all()
-    teachers = []
-    for i in res:
-        temp = {"profesor_id":i[0],"usuario_username":i[3],"usuario_email":i[4],"profesor_experiencia":i[1],"profesor_curso":i[2]}
-        teachers.append(temp)
-    return jsonify(teachers)
 
 @app.route('/taught/<student_name>', methods=['GET'])
 def route_clase_student_name(student_name):
@@ -163,29 +126,6 @@ def route_clase_teacher_name(teacher_name):
             temp = {"ensenanza_id":i[0],"ensenanza_clase_aprobada":i[1],"ensenanza_url":i[2],"ensenanza_fecha":i[3],"estudiante_username":i[5]}
             clase.append(temp)
         return jsonify(clase)
-
-def insert_student(student_data):
-    #Inserta el nuevo estudiante en la tabla usuario
-    user = Usuario(username=student_data['username'],email=student_data['email'],password=student_data['password'])
-    db.session.add(user)
-    db.session.commit()
-    
-    #Inserta el nuevo estudiante en la tabla estudiante
-    student = Estudiante(usuario_id=user.id)
-    db.session.add(student)
-    db.session.commit()
-    return "SUCCESS"
-    
-def insert_teacher(teacher_data):
-
-    user = Usuario(username=teacher_data['username'],email=teacher_data['email'],password=teacher_data['password'])
-    db.session.add(user)
-    db.session.commit()
-
-    teacher = Profesor(usuario_id=user.id,experiencia=teacher_data['experiencia'],curso=teacher_data["curso"])
-    db.session.add(teacher)
-    db.session.commit()
-    return "SUCCESS"
 
 def insert_clase(data):
     student = Estudiante.query.join(Usuario, Estudiante.usuario_id==Usuario.id).filter(Usuario.username == data["estudiante_username"]).add_columns(Estudiante.id).first()
@@ -223,22 +163,6 @@ def delete_clase(clase_id):
     db.session.commit()
     return 'SUCCESS'
 
-def get_students():
-    res = db.session.query(Estudiante.id,Usuario.username,Usuario.email,Usuario.password).filter(Estudiante.usuario_id==Usuario.id).all()
-    students = []
-    for i in res:
-        temp = {"estudiante_id":i[0],"usuario_username":i[1],"usuario_email":i[2],"usuario_password":i[3]}
-        students.append(temp)
-    return jsonify(students)
-
-def get_teachers():
-    res = db.session.query(Profesor.id,Usuario.username,Usuario.email,Usuario.password).filter(Profesor.usuario_id==Usuario.id).all()
-    teachers = []
-    for i in res:
-        temp = {"profesor_id":i[0],"usuario_username":i[1],"usuario_email":i[2],"usuario_password":i[3]}
-        teachers.append(temp)
-    return jsonify(teachers)
-
 #def delete_student(student_id):
 #    student = Estudiante.query.get_or_404(student_id)
 #    db.session.delete(student)
@@ -252,4 +176,4 @@ def test_connection():
     return "Conection Working"
 
 if __name__ == "__main__":
-    app.run(host='0.0.0.0', port=8000, debug=False)
+    app.run(host='0.0.0.0', port=8003, debug=False)
